Don't hide chat message input field in SC thread modal.
The message input in the SC thread modal would be hidden by default,
just like the main chat message input field at the bottom of the chat window.
However the SC thread modal is not transparent anyway so there's no point in doing so.
This makes it so that the hiding CSS does not hide the message input in the SC thread modal.